Airline-Tickets/requirements.txt
astrapy
python-dotenv
cassandra-driver
